--- MyMoney.Web/ClientApp/src/app/helpers/jwt.interceptor.ts
@@ -1,24 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
-import { Observable } from 'rxjs';
-
-import { AuthenticationService } from '../authentication.service';
-
-@Injectable()
-export class JwtInterceptor implements HttpInterceptor {
-   constructor(private authenticationService: AuthenticationService) { }
-
-   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-      // add authorization header with jwt token if available
-      const currentUser = this.authenticationService.currentUserValue;
-      if (currentUser && currentUser.token) {
-         request = request.clone({
-            setHeaders: {
-               Authorization: `Bearer ${currentUser.token}`
-            }
-         });
-      }
-
-      return next.handle(request);
-   }
-}
