--- MyMoney.Web/ClientApp/src/app/models/budget.list.response.ts
@@ -1,5 +1,0 @@
-import { BudgetModel } from './budget.model';
-
-export interface BudgetListResponse {
-   budgets: Array<BudgetModel>;
-}
