--- MyMoney.Web/ClientApp/src/app/models/transaction.list.response.ts
@@ -1,5 +1,0 @@
-import { TransactionModel } from './transaction.model';
-
-export interface TransactionListResponse {
-   transactions: Array<TransactionModel>;
-}
