+++ MyMoney.Web/ClientApp/src/app/pages/home/home.component.html
