+++ MyMoney.Web/ClientApp/src/app/pages/home/home.component.ts
@@ -1,0 +1,20 @@
+import { Component } from '@angular/core';
+import { Router } from '@angular/router';
+
+import { AuthenticationService } from '../../shared/services';
+
+@Component({
+   templateUrl: './home.component.html',
+})
+export class HomeComponent {
+
+   constructor(
+      private readonly authenticationService: AuthenticationService,
+      private readonly router: Router
+   ) {
+      // redirect to home if already logged in
+      if (!this.authenticationService.isLoggedIn) {
+         this.router.navigate(['/login']);
+      }
+   }
+}
