+++ MyMoney.Web/ClientApp/src/app/pages/index.ts
@@ -1,0 +1,9 @@
+export * from './home/home.component';
+export * from './login/login.component';
+export * from './register/register.component';
+export * from './transactions/add/add.transactions.component';
+export * from './transactions/transactions.component';
+export * from './budgets/budgets.component';
+export * from './budgets/add/add.budgets.component';
+export * from './transactions/edit/edit.transactions.component';
+export * from './budgets/edit/edit.budgets.component';
