+++ MyMoney.Web/ClientApp/src/app/pages/register/register.component.html
