+++ MyMoney.Web/ClientApp/src/app/pages/transactions/add/add.transactions.component.html
