+++ MyMoney.Web/ClientApp/src/app/pages/transactions/edit/edit.transactions.component.html
