+++ MyMoney.Web/ClientApp/src/app/pages/transactions/transactions.component.html
