+++ MyMoney.Web/ClientApp/src/app/shared/classes/index.ts
@@ -1,0 +1,3 @@
+export * from './jwt-interceptor.class';
+export * from './view-models';
+
