+++ MyMoney.Web/ClientApp/src/app/shared/classes/view-models/budget-view-model.class.ts
@@ -1,0 +1,19 @@
+import { BudgetModel } from '../../interfaces/budget-model.interface';
+
+export class BudgetViewModel {
+   id: Number;
+   amount: string;
+   notes: string;
+   monthId: string;
+   name: string;
+   remaining: string;
+
+   constructor(model: BudgetModel) {
+      this.monthId = model.monthId;
+      this.name = model.name;
+      this.notes = model.notes;
+      this.amount = '£' + model.amount;
+      this.id = model.id;
+      this.remaining = '£' + model.remaining;
+   }
+}
