+++ MyMoney.Web/ClientApp/src/app/shared/classes/view-models/index.ts
@@ -1,0 +1,5 @@
+export * from './budget-view-model.class';
+export * from './date-range-view-model.class';
+export * from './transaction-view-model.class';
+
+
