+++ MyMoney.Web/ClientApp/src/app/shared/classes/view-models/transaction-view-model.class.ts
@@ -1,0 +1,15 @@
+import { TransactionModel } from '../../interfaces/transaction-model.interface';
+
+export class TransactionViewModel {
+   date: string;
+   description: string;
+   amount: string;
+   id: Number;
+
+   constructor(model: TransactionModel) {
+      this.date = model.date;
+      this.description = model.description;
+      this.amount = '£' + model.amount;
+      this.id = model.id;
+   }
+}
