+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/budget-list-response.interface.ts
@@ -1,0 +1,5 @@
+import { BudgetModel } from './budget-model.interface';
+
+export interface BudgetListResponse {
+   budgets: Array<BudgetModel>;
+}
