+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/date-range-model.interface.ts
