+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/delete-response.interface.ts
