+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/index.ts
@@ -1,0 +1,11 @@
+export * from './budget-list-response.interface';
+export * from './budget-model.interface';
+export * from './budget-request.interface';
+export * from './date-range-model.interface';
+export * from './delete-response.interface';
+export * from './login-response.interface';
+export * from './register-request.interface';
+export * from './transaction-list-response.interface';
+export * from './transaction-model.interface';
+export * from './update.response.interface';
+export * from './user.interface';
