+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/login-response.interface.ts
