+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/register-request.interface.ts
