+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/transaction-list-response.interface.ts
@@ -1,0 +1,5 @@
+import { TransactionModel } from './transaction-model.interface';
+
+export interface TransactionListResponse {
+   transactions: Array<TransactionModel>;
+}
