+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/transaction-model.interface.ts
