+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/update.response.interface.ts
