+++ MyMoney.Web/ClientApp/src/app/shared/interfaces/user.interface.ts
