+++ MyMoney.Web/ClientApp/src/app/shared/services/index.ts
@@ -1,0 +1,1 @@
+export * from './authentication-service.service';
